+++ app/scss/style.scss
@@ -220,7 +220,7 @@ hr {
   overflow: hidden;
 }
 
-#timeNodeStatistics.table {
+#timeNodeDashboard.table {
   & tbody tr td {
     padding: 0.75rem;
     border-top: 1px solid #dee2e6;
@@ -373,3 +373,4 @@ hr {
   max-width: 75%;
   min-width: 140px;
 }
+
